+++ src/main/java/org/openpnp/machine/reference/ReferenceHead.java
@@ -26,7 +26,9 @@ import javax.swing.Action;
 import org.openpnp.ConfigurationListener;
 import org.openpnp.gui.support.PropertySheetWizardAdapter;
 import org.openpnp.gui.support.Wizard;
+import org.openpnp.machine.reference.psh.ActuatorsPropertySheetHolder;
 import org.openpnp.machine.reference.psh.CamerasPropertySheetHolder;
+import org.openpnp.machine.reference.psh.NozzlesPropertySheetHolder;
 import org.openpnp.machine.reference.wizards.ReferenceHeadConfigurationWizard;
 import org.openpnp.model.Configuration;
 import org.openpnp.spi.PropertySheetHolder;
@@ -70,9 +72,9 @@ public class ReferenceHead extends AbstractHead {
     @Override
     public PropertySheetHolder[] getChildPropertySheetHolders() {
         ArrayList<PropertySheetHolder> children = new ArrayList<>();
-        children.add(new SimplePropertySheetHolder("Nozzles", getNozzles()));
+        children.add(new NozzlesPropertySheetHolder(this, "Nozzles", getNozzles(), null));
         children.add(new CamerasPropertySheetHolder(this, "Cameras", getCameras(), null));
-        children.add(new SimplePropertySheetHolder("Actuators", getActuators()));
+        children.add(new ActuatorsPropertySheetHolder(this, "Actuators", getActuators(), null));
         children.add(new SimplePropertySheetHolder("Paste Dispensers", getPasteDispensers()));
         return children.toArray(new PropertySheetHolder[] {});
     }