+++ src/main/java/org/openpnp/machine/reference/ReferenceMachine.java
@@ -40,14 +40,17 @@ import org.openpnp.machine.reference.feeder.ReferenceDragFeeder;
 import org.openpnp.machine.reference.feeder.ReferenceStripFeeder;
 import org.openpnp.machine.reference.feeder.ReferenceTrayFeeder;
 import org.openpnp.machine.reference.feeder.ReferenceTubeFeeder;
+import org.openpnp.machine.reference.psh.ActuatorsPropertySheetHolder;
 import org.openpnp.machine.reference.psh.CamerasPropertySheetHolder;
 import org.openpnp.machine.reference.vision.ReferenceBottomVision;
 import org.openpnp.machine.reference.vision.ReferenceFiducialLocator;
 import org.openpnp.machine.reference.wizards.ReferenceMachineConfigurationWizard;
+import org.openpnp.spi.Actuator;
 import org.openpnp.spi.Camera;
 import org.openpnp.spi.Feeder;
 import org.openpnp.spi.FiducialLocator;
 import org.openpnp.spi.Head;
+import org.openpnp.spi.Nozzle;
 import org.openpnp.spi.PartAlignment;
 import org.openpnp.spi.PasteDispenseJobProcessor;
 import org.openpnp.spi.PnpJobProcessor;
@@ -143,7 +146,7 @@ public class ReferenceMachine extends AbstractMachine {
         children.add(new SimplePropertySheetHolder("Feeders", getFeeders()));
         children.add(new SimplePropertySheetHolder("Heads", getHeads()));
         children.add(new CamerasPropertySheetHolder(null, "Cameras", getCameras(), null));
-        children.add(new SimplePropertySheetHolder("Actuators", getActuators()));
+        children.add(new ActuatorsPropertySheetHolder(null, "Actuators", getActuators(), null));
         children.add(
                 new SimplePropertySheetHolder("Driver", Collections.singletonList(getDriver())));
         children.add(new SimplePropertySheetHolder("Job Processors",
@@ -191,6 +194,20 @@ public class ReferenceMachine extends AbstractMachine {
         l.add(OnvifIPCamera.class);
         l.add(ImageCamera.class);
         l.add(SimulatedUpCamera.class);
+        return l;
+    }
+    
+    @Override
+    public List<Class<? extends Nozzle>> getCompatibleNozzleClasses() {
+        List<Class<? extends Nozzle>> l = new ArrayList<>();
+        l.add(ReferenceNozzle.class);
+        return l;
+    }
+
+    @Override
+    public List<Class<? extends Actuator>> getCompatibleActuatorClasses() {
+        List<Class<? extends Actuator>> l = new ArrayList<>();
+        l.add(ReferenceActuator.class);
         return l;
     }