+++ src/main/java/org/openpnp/spi/base/AbstractNozzle.java
@@ -6,6 +6,7 @@ import java.util.List;
 import javax.swing.Icon;
 
 import org.openpnp.gui.support.Icons;
+import org.openpnp.model.AbstractModelObject;
 import org.openpnp.model.Configuration;
 import org.openpnp.model.Part;
 import org.openpnp.spi.Head;
@@ -15,7 +16,7 @@ import org.openpnp.util.IdentifiableList;
 import org.simpleframework.xml.Attribute;
 import org.simpleframework.xml.ElementList;
 
-public abstract class AbstractNozzle implements Nozzle {
+public abstract class AbstractNozzle extends AbstractModelObject implements Nozzle {
     @ElementList(required = false)
     protected IdentifiableList<NozzleTip> nozzleTips = new IdentifiableList<>();
 
@@ -62,6 +63,7 @@ public abstract class AbstractNozzle implements Nozzle {
     @Override
     public void setName(String name) {
         this.name = name;
+        firePropertyChange("name", null, name);
     }
     
     @Override